Engine patches/bugfixes from ITGmania (2021-2024)
Spans about https://github.com/itgmania/itgmania/commits/v0.5.1/?after=e2a48eb7267103b28436760a839819a2ec006e7d+174 to 9737482f7c50dccd52ea36aecad7127fbbf942f4. This is all the MIT licensed commits from ITGmania that do not include ITG specific changes. In other words, this is just the engine side improvements and fixed build system commits separated out for use in  plain SM5.

Notable changes:
Built-in network functionality via mbedtls (replaces the old SMOnlineScreen).
Updated/fixed CMake integration
Working build system for all modern OS's
Reload new songs from within the song select screen
The mine fix applied (courtesy of DinsFire64)
Held misses tracked in the engine for pad debugging
Fixed overlapping hold bug
Per-player visual delay
Per-player disabling of timing windows
New preference to control note render ordering
Increased the Stats.xml file size limit to 100MB
TONS of code health fixes
C++11 compliance
Various fixes for modern compilers and build systems
libusb/hidapi integration
Linux PacDrive support

+++ src/ScreenDimensions.cpp
@@ -5,11 +5,13 @@
 #include "LuaManager.h"
 #include "ThemeMetric.h"
 
+#include <cmath>
+
 static ThemeMetric<float> THEME_SCREEN_WIDTH("Common","ScreenWidth");
 static ThemeMetric<float> THEME_SCREEN_HEIGHT("Common","ScreenHeight");
 
 /* The theme's logical resolution specifies the minimum screen width and
- * the minimum screen height with a 4:3 aspect ratio. Scale just one 
+ * the minimum screen height with a 4:3 aspect ratio. Scale just one
  * of the dimensions up to meet the requested aspect ratio. */
 
 /* The theme resolution isn't necessarily 4:3; a natively widescreen
@@ -24,8 +26,8 @@ float ScreenDimensions::GetThemeAspectRatio()
 	return THEME_NATIVE_ASPECT;
 }
 
-/* ceilf was originally lrintf. However, lrintf causes odd resolutions like
- * 639x480 (4:3) and 853x480 (16:9). ceilf gives the correct values of 640x480
+/* ceil was originally lrint. However, lrint causes odd resolutions like
+ * 639x480 (4:3) and 853x480 (16:9). ceil gives the correct values of 640x480
  * and 854x480 (should really be 852 so that SCREEN_CENTER_X == 426 and not 427)
  * respectively. -aj */
 float ScreenDimensions::GetScreenWidth()
@@ -35,9 +37,9 @@ float ScreenDimensions::GetScreenWidth()
 	if( fAspect > THEME_NATIVE_ASPECT )
 		fScale = fAspect / THEME_NATIVE_ASPECT;
 	ASSERT( fScale >= 1 );
-	// ceilf causes the width to come out odd when it shouldn't.
+	// ceil causes the width to come out odd when it shouldn't.
 	// 576 * 1.7778 = 1024.0128, which is rounded to 1025. -Kyz
-	int width= (int)ceilf(THEME_SCREEN_WIDTH * fScale);
+	int width= std::ceil(THEME_SCREEN_WIDTH * fScale);
 	width-= width % 2;
 	return (float)width;
 }
@@ -49,7 +51,7 @@ float ScreenDimensions::GetScreenHeight()
 	if( fAspect < THEME_NATIVE_ASPECT )
 		fScale = THEME_NATIVE_ASPECT / fAspect;
 	ASSERT( fScale >= 1 );
-	return (float) ceilf(THEME_SCREEN_HEIGHT * fScale);
+	return std::ceil(THEME_SCREEN_HEIGHT * fScale);
 }
 
 void ScreenDimensions::ReloadScreenDimensions()
@@ -77,7 +79,7 @@ LuaFunction( GetThemeAspectRatio,	ScreenDimensions::GetThemeAspectRatio() );
 /*
  * (c) 2001-2002 Chris Danford
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -87,7 +89,7 @@ LuaFunction( GetThemeAspectRatio,	ScreenDimensions::GetThemeAspectRatio() );
  * copyright notice(s) and this permission notice appear in all copies of
  * the Software and that both the above copyright notice(s) and this
  * permission notice appear in supporting documentation.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
  * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT OF
